pyproject.toml
[tool.poetry]
name = "trading diary"
version = "0.1.0"
description = ""
authors = ["Zinchenko Nikolai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
gunicorn = "^20.1.0"
django-filer = "^2.2.4"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
environs = "^9.5.0"
django-admin-rangefilter = "^0.10.0"


[tool.poetry.group.tests.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
